backend/requirements.txt
fastapi
uvicorn
alpaca-trade-api
polygon-api-client
tavily-python
python-dotenv
sqlalchemy
greenlet
aiofiles
aiosqlite
httpx
yfinance
openai
